Too close for missiles, switching to guns.  Or, retain/release is too hard; switching to garbage collection.
This change introduces a simple garbage collector.  All of the
retain/release/refcount code is tossed out.  The grabage collector could
use some help to make it more efficient.  But, it should be fine for a
wile.

More test code is needed.  Also, slightly changed the print code to not
always print out all the meta-information (object address, for example).