tyForth's biggest change to-date!!
The mechanism by which strings are executed has completely changed: they
are compiled into a temporary word and then that word is executed!

constants work in the three-step dosi-do manner that they do.  (See the
source for comments on that process.)

There's code that needs to be ripped out now that's no longer needed.  But,
I'll take care of that soon enough.  I just want to commit this and push it
out there to deal with later.

This is a huge change.  Compiling words are present!  There are still bugs.
But it's pretty good.